Added an option to record event 'life time'
The 'life time' of an event, as I define it, is the time from the head node finishes processing an event request, to the time it is inserted into commit request queue.

To enable the recording, set EVENT_LIFE_TIME parameter to 1 (default 0). It uses Accumulator to show the accumulated lifetime, so be sure to turn on accumulator feature:

MACE_LOG_AUTO_SELECTORS = "Accumulator::EVENT_LIFE_TIME"
and
MACE_LOG_ACCUMULATOR =1000

The life is accumulated in microseconds.