Removed unused/redundant code in perl compiler, and mark several internal functions in ContextService as private.
The code is much much much cleaner than before. What's more it is now possible to write a context application purely in C++ with a few lines of code.

Compilation speed is improved as well: It takes only 2/3 of time to compile an application (testrun/contextrun) The time reduction comes from the reduced
generated code: it reduces the time the perl compiler has to generate the code, and reduces the time g++ has to spend to compile the code. Compile time
is typically proportional to the length of generated code.

--HG--
branch : chuangw/multicompiler