The 1000th changeset is pushed up to repository!!
A huge milestone to this project. Since December 2011 we started this project, it's been 16 months already.
This is a huge huge project that relies on so many people's contribution.

Statistics:

In these 1000 changesets, we made 40401 lines of code of change in 351 files.

That contains 5700 lines of perl, 15700 lines of C++, 17900 lines of Mace code and some shell scripts.

Contribution was made by 7 members in 16 months.

--HG--
branch : chuangw/multicompiler