Added a test case (or a benchmark in a sense) to show the interaction performance of two logical nodes.
In TestCase3, a client sends a number of messages in multiple contexts simultanouesly to a server.

Updated the code generator to correctly parse the context annotation of a message delivery transition , if the context is of multi-dimensional.

Attempted to use MaceKey as the context index, (through specializing the boost::lexical_cast<> function) but failed. Will find a way to do this right.

--HG--
branch : chuangw/multicompiler