Simply createContextObject() interface . Reduce the number of exposed API to the services.
Right now, in ContextService, only one member variable and 15 methods are exposed the derived service classes.

That should make things easier to work with. On the other hand, there are 78 private methods and 11 private variables that
previously were all exposed to the derived service classes.

This code is througly tested and I am going to merge it to default branch.

--HG--
branch : chuangw/multicompiler